REFACTOR : split and rename tests
